Fix: differentiate attribute overrides in convert proielxml
This fixes a bug in which all alignment-ids for all tokens and sentences
are set to the alignment-id of the div if alignment inference is
enabled. This happens because the `override` hash, which is used to
override attribute values based on filtering operations etc., is shared
between divs, sentences and tokens. On the other hand, the `override`
cannot be local to, for example, the token processing block as the
`antecedent_id` override is actually shared between all tokens in a div.

This is a workaround for this issue; we need a more explicit system for
tracking the context of overrides.