Update test data and expectation to match stricted conversion checks.
One of the previous commits that updated the CoNLL-U conversion code
introduced a check on the linear sequence of elements with the `conj`
relation. This broke once of the Cucumber acceptance tests since the
test data was not updated at same time.

I've fixed this by updating the test data using the most recent version
of the relevant text:

```
bundle exec proiel convert proielxml \
 --remove-not-reviewed \
../syntacticus-treebank-data-dev/caes-gal.xml > features/caes-gal-convertible.xml
```